feat(commonware-node): define peers using DNS names (#193)
Changes the peers entry in the commonware config to take a map of `(public-key, String)` instead of `(public-key, SocketAddr)`. The values will now be resolved lazily to socket addresses at runtime.

This implementation is very strict: if resolution for even one peer fails (including if it does not have an address), then the entire node will refuse to start instead of just ignoring that peer.

We can loosen that later.

Closes #194