fix: update size fn for tempo transaction (#1298)
Fixes - [Zellic Issue
](https://tempoxyz.slack.com/archives/C09QPJXBPE3/p1765193916423139)

It should be okay to not hardfork gate this, as this size function is
not used in any critical places.

It is only in TempoPooledTransaction I believe, where it's used to limit
the size of the TxPool in memory, so this minor underestimation should
not be critical.

---------

Co-authored-by: Matthias Seitz <matthias.seitz@outlook.de>